+++ filament/backend/src/metal/MetalDriverFactory.h
@@ -17,6 +17,8 @@
 #ifndef TNT_FILAMENT_DRIVER_METALDRIVERFACTORY_H
 #define TNT_FILAMENT_DRIVER_METALDRIVERFACTORY_H
 
+#include <backend/Platform.h>
+
 namespace filament {
 namespace backend {
 class MetalPlatform;
@@ -24,7 +26,7 @@ class Driver;
 
 class MetalDriverFactory {
 public:
-    static Driver* create(MetalPlatform* platform);
+    static Driver* create(MetalPlatform* platform, const Platform::DriverConfig& driverConfig);
 };
 
 } // namespace backend