+++ filament/backend/src/opengl/OpenGLDriverFactory.h
@@ -17,6 +17,8 @@
 #ifndef TNT_FILAMENT_BACKEND_OPENGL_OPENGLDRIVERFACTORY_H
 #define TNT_FILAMENT_BACKEND_OPENGL_OPENGLDRIVERFACTORY_H
 
+#include <backend/Platform.h>
+
 namespace filament::backend {
 
 class OpenGLPlatform;
@@ -24,7 +26,8 @@ class Driver;
 
 class OpenGLDriverFactory {
 public:
-    static Driver* create(OpenGLPlatform* platform, void* sharedGLContext) noexcept;
+    static Driver* create(OpenGLPlatform* platform, void* sharedGLContext,
+            const Platform::DriverConfig& driverConfig) noexcept;
 };
 
 } // namespace filament::backend