+++ filament/backend/src/opengl/platforms/PlatformWebGL.h
@@ -31,7 +31,7 @@ namespace filament::backend {
 class PlatformWebGL final : public OpenGLPlatform {
 public:
 
-    Driver* createDriver(void* const sharedGLContext) noexcept override;
+    Driver* createDriver(void* const sharedGLContext, const Platform::DriverConfig& driverConfig) noexcept override;
     void terminate() noexcept override;
 
     SwapChain* createSwapChain(void* nativewindow, uint64_t& flags) noexcept final override;