+++ filament/backend/src/vulkan/PlatformVkWindows.cpp
@@ -27,10 +27,10 @@ using namespace bluevk;
 
 namespace filament::backend {
 
-Driver* PlatformVkWindows::createDriver(void* const sharedContext) noexcept {
+Driver* PlatformVkWindows::createDriver(void* const sharedContext, const Platform::DriverConfig& driverConfig) noexcept {
     ASSERT_PRECONDITION(sharedContext == nullptr, "Vulkan does not support shared contexts.");
     const char* requiredInstanceExtensions[] = { "VK_KHR_win32_surface" };
-    return VulkanDriverFactory::create(this, requiredInstanceExtensions, 1);
+    return VulkanDriverFactory::create(this, requiredInstanceExtensions, 1, driverConfig);
 }
 
 void* PlatformVkWindows::createVkSurfaceKHR(void* nativeWindow, void* instance, uint64_t flags) noexcept {