+++ filament/backend/src/vulkan/VulkanDriverFactory.h
@@ -17,6 +17,8 @@
 #ifndef TNT_FILAMENT_BACKEND_VULKANDRIVERFACTORY_H
 #define TNT_FILAMENT_BACKEND_VULKANDRIVERFACTORY_H
 
+#include <backend/Platform.h>
+
 #include <stdint.h>
 
 namespace filament::backend {
@@ -27,7 +29,8 @@ class Driver;
 class VulkanDriverFactory {
 public:
     static Driver* create(VulkanPlatform* platform,
-            const char* const* ppEnabledExtensions, uint32_t enabledExtensionCount) noexcept;
+            const char* const* ppEnabledExtensions, uint32_t enabledExtensionCount,
+            const Platform::DriverConfig& driverConfig) noexcept;
 };
 
 } // namespace filament::backend